move cache key validation
