+++ app/routers/parser.py
@@ -19,6 +19,7 @@ from app.models.parser import (
     ListParsersResponse,
 )
 from app.redis_cache import TRedisPlugin
+from app.redis_cache._cache import _validate_cache_key
 from app.routers.session import _update_session, _update_session_list_item
 
 if TYPE_CHECKING:  # pragma: no cover
@@ -32,19 +33,6 @@ ROUTER = APIRouter(
     tags=["parser"],
     responses={status.HTTP_404_NOT_FOUND: {"model": HTTPNotFoundError}},
 )
-
-
-async def _validate_cache_key(cache: TRedisPlugin, key: str, key_type: str) -> None:
-    """Validate if a key exists in cache and is of expected type (str or bytes)."""
-    if not await cache.exists(key):
-        raise httpexception_404_item_id_does_not_exist(key, key_type)
-
-    cache_value = await cache.get(key)
-    if not isinstance(cache_value, (str, bytes)):
-        raise TypeError(
-            f"Expected cache value of {key} to be a string or bytes, "
-            f"found it to be of type: `{type(cache_value)!r}`."
-        )
 
 
 # Create parser
